chore(api): restructure to replicate normal dir structure
ref: #319 #320 nofusscomputing/centurion_erp_ui#1
